refactor: move experimental circuit as default (tests not updated)
