+++ anchor/programs/darklake/Cargo.toml
@@ -22,3 +22,4 @@ anchor-spl = "0.30.1"
 groth16-solana = "0.0.3"
 solana-program = "2.0.1"
 mpl-token-metadata = "4.1.2"
+spl-math = "0.3"
