+++ anchor/programs/darklake/src/constants/mod.rs
@@ -1,5 +1,3 @@
 pub mod verifying_key;
-pub mod verifying_key_exp;
 
 pub use verifying_key::VERIFYINGKEY;
-pub use verifying_key_exp::VERIFYINGKEYEXP;
