+++ anchor/programs/darklake/src/constants/verifying_key.rs
@@ -1,18 +1,18 @@
 use groth16_solana::groth16::Groth16Verifyingkey;
 
 pub const VERIFYINGKEY: Groth16Verifyingkey =  Groth16Verifyingkey {
-	nr_pubinputs: 4,
+	nr_pubinputs: 5,
 
 	vk_alpha_g1: [
-		22,184,82,43,251,173,24,249,220,191,156,184,102,88,56,109,52,50,24,144,181,98,173,106,144,8,174,121,20,34,77,110,
-		32,224,246,194,237,231,247,227,244,158,248,4,245,117,28,240,222,31,119,229,171,191,33,243,54,215,119,189,222,91,104,142,
+		39,173,3,22,55,111,32,141,9,85,21,8,48,39,24,59,235,141,243,74,21,130,185,94,209,174,60,77,125,154,106,44,
+		14,192,150,8,92,51,136,180,189,92,162,145,148,0,17,146,117,62,177,169,88,152,4,181,201,18,176,120,95,229,105,154,
 	],
 
 	vk_beta_g2: [
-		33,130,157,216,18,230,178,16,139,18,132,142,132,12,97,43,205,184,225,81,70,124,13,196,107,231,252,160,54,19,19,151,
-		17,4,37,201,121,232,161,121,33,24,239,168,211,250,223,53,232,1,237,91,113,80,92,220,140,151,235,169,79,22,56,229,
-		31,17,68,3,243,143,3,64,149,90,37,69,119,66,50,174,79,40,188,20,61,227,102,49,220,173,144,186,186,220,173,37,
-		18,4,54,241,6,20,10,176,100,41,38,130,157,21,153,145,57,24,18,161,107,246,141,54,223,169,174,93,10,97,120,115,
+		8,137,88,168,140,170,115,83,207,230,34,118,172,25,29,107,150,221,65,207,169,246,42,246,119,221,91,248,181,18,149,184,
+		36,85,82,195,72,81,198,58,35,53,247,176,67,128,41,168,62,78,146,65,115,79,32,188,44,27,79,74,86,173,84,77,
+		19,102,102,239,208,205,239,54,66,186,87,147,1,147,86,39,229,76,45,231,102,137,233,137,98,173,199,124,30,63,76,148,
+		42,106,120,138,62,124,42,148,84,17,86,92,112,213,148,39,16,48,126,195,166,255,179,32,0,170,149,99,17,147,249,73,
 	],
 
 	vk_gamme_g2: [
@@ -23,28 +23,32 @@ pub const VERIFYINGKEY: Groth16Verifyingkey =  Groth16Verifyingkey {
 	],
 
 	vk_delta_g2: [
-		27,108,36,220,246,250,217,224,247,233,196,50,34,216,229,12,200,209,213,72,143,212,160,250,61,9,230,177,118,65,161,199,
-		28,97,132,229,16,114,222,53,231,129,88,179,133,247,244,184,112,122,247,14,171,132,237,191,122,50,86,227,76,31,185,250,
-		14,106,112,194,113,38,64,172,46,164,2,22,132,85,97,174,128,237,79,229,79,86,12,156,186,163,234,105,137,12,79,118,
-		26,24,158,99,117,187,181,252,184,70,169,98,106,29,13,188,99,99,69,245,2,101,12,175,144,237,190,224,44,163,122,108,
+		20,82,32,24,191,48,7,155,2,116,206,219,22,227,161,110,174,1,186,62,32,229,184,192,112,41,46,152,120,201,111,52,
+		19,170,107,51,66,113,23,179,171,188,220,184,250,211,249,214,191,66,251,175,193,141,108,34,252,48,53,241,194,179,141,190,
+		13,64,244,20,180,145,51,198,177,4,98,247,147,103,98,231,188,74,168,254,187,203,89,17,5,101,24,141,98,23,163,50,
+		16,114,187,153,115,69,54,226,122,54,234,114,69,91,99,159,85,54,207,35,202,119,143,181,91,165,24,227,4,77,144,112,
 	],
 
 	vk_ic: &[
 		[
-			31,16,224,153,65,115,132,119,160,193,212,228,141,189,25,224,74,251,241,113,94,187,254,240,25,137,35,160,109,216,26,141,
-			47,100,50,11,75,238,147,51,221,70,133,128,15,129,17,224,209,117,195,193,76,65,133,221,61,26,10,223,136,148,167,89,
+			20,117,178,144,199,86,40,28,139,72,149,150,221,34,220,185,112,196,247,204,36,86,0,111,2,39,252,64,163,136,158,72,
+			14,116,172,32,18,129,193,47,179,128,153,220,86,97,240,34,171,181,104,5,72,194,175,108,20,184,146,61,171,238,127,250,
 		],
 		[
-			0,95,135,5,80,75,231,132,60,152,114,39,187,3,254,195,46,254,145,192,133,225,186,184,9,249,51,32,132,153,108,235,
-			9,15,70,57,69,0,15,36,88,26,31,61,30,215,218,148,58,112,35,101,94,228,144,171,113,210,90,118,20,25,16,209,
+			32,83,181,76,227,121,250,47,99,98,192,148,197,165,111,76,36,158,155,160,246,115,203,238,253,254,139,250,176,51,162,229,
+			26,203,227,24,200,114,78,250,92,69,180,67,48,84,214,122,58,41,134,87,194,168,123,165,103,228,149,242,190,208,204,40,
 		],
 		[
-			6,218,135,224,63,93,74,202,246,250,152,209,160,227,58,148,136,199,196,42,141,194,152,167,90,182,18,109,218,136,155,84,
-			19,10,142,87,192,88,100,68,78,214,187,44,21,43,161,14,110,58,94,87,196,135,92,245,219,211,74,236,114,61,178,18,
+			42,185,115,152,171,110,26,235,67,184,206,136,164,59,136,168,243,217,252,214,161,247,124,225,28,54,89,216,180,178,156,212,
+			37,219,158,110,208,228,240,100,139,45,98,39,5,120,91,68,178,248,209,142,68,44,161,99,67,228,197,219,137,238,36,106,
 		],
 		[
-			23,176,98,99,55,89,82,250,198,141,16,221,114,243,139,24,65,165,216,193,21,239,27,41,52,230,249,100,74,211,81,227,
-			3,28,106,192,178,16,184,99,16,24,118,106,93,203,42,159,94,154,128,145,40,255,128,185,252,40,194,201,106,233,205,50,
+			10,61,208,60,178,139,175,144,164,251,164,51,41,242,227,193,62,91,83,250,34,163,155,251,97,92,68,187,170,230,204,186,
+			15,74,29,129,199,37,101,163,127,57,111,6,135,106,61,37,67,103,196,228,27,134,237,126,248,183,26,110,106,124,86,132,
+		],
+		[
+			22,163,116,25,68,151,95,1,98,83,99,172,107,98,128,76,213,51,155,160,104,61,216,3,114,144,56,172,146,235,85,124,
+			19,251,92,184,124,129,6,66,241,98,69,167,26,12,193,77,162,141,144,139,106,118,227,121,247,220,23,4,243,133,14,113,
 		],
 	]
 };
