--- anchor/programs/darklake/src/constants/verifying_key_exp.rs
@@ -1,54 +1,0 @@
-use groth16_solana::groth16::Groth16Verifyingkey;
-
-pub const VERIFYINGKEYEXP: Groth16Verifyingkey =  Groth16Verifyingkey {
-	nr_pubinputs: 5,
-
-	vk_alpha_g1: [
-		39,173,3,22,55,111,32,141,9,85,21,8,48,39,24,59,235,141,243,74,21,130,185,94,209,174,60,77,125,154,106,44,
-		14,192,150,8,92,51,136,180,189,92,162,145,148,0,17,146,117,62,177,169,88,152,4,181,201,18,176,120,95,229,105,154,
-	],
-
-	vk_beta_g2: [
-		8,137,88,168,140,170,115,83,207,230,34,118,172,25,29,107,150,221,65,207,169,246,42,246,119,221,91,248,181,18,149,184,
-		36,85,82,195,72,81,198,58,35,53,247,176,67,128,41,168,62,78,146,65,115,79,32,188,44,27,79,74,86,173,84,77,
-		19,102,102,239,208,205,239,54,66,186,87,147,1,147,86,39,229,76,45,231,102,137,233,137,98,173,199,124,30,63,76,148,
-		42,106,120,138,62,124,42,148,84,17,86,92,112,213,148,39,16,48,126,195,166,255,179,32,0,170,149,99,17,147,249,73,
-	],
-
-	vk_gamme_g2: [
-		25,142,147,147,146,13,72,58,114,96,191,183,49,251,93,37,241,170,73,51,53,169,231,18,151,228,133,183,174,243,18,194,
-		24,0,222,239,18,31,30,118,66,106,0,102,94,92,68,121,103,67,34,212,247,94,218,221,70,222,189,92,217,146,246,237,
-		9,6,137,208,88,95,240,117,236,158,153,173,105,12,51,149,188,75,49,51,112,179,142,243,85,172,218,220,209,34,151,91,
-		18,200,94,165,219,140,109,235,74,171,113,128,141,203,64,143,227,209,231,105,12,67,211,123,76,230,204,1,102,250,125,170,
-	],
-
-	vk_delta_g2: [
-		20,82,32,24,191,48,7,155,2,116,206,219,22,227,161,110,174,1,186,62,32,229,184,192,112,41,46,152,120,201,111,52,
-		19,170,107,51,66,113,23,179,171,188,220,184,250,211,249,214,191,66,251,175,193,141,108,34,252,48,53,241,194,179,141,190,
-		13,64,244,20,180,145,51,198,177,4,98,247,147,103,98,231,188,74,168,254,187,203,89,17,5,101,24,141,98,23,163,50,
-		16,114,187,153,115,69,54,226,122,54,234,114,69,91,99,159,85,54,207,35,202,119,143,181,91,165,24,227,4,77,144,112,
-	],
-
-	vk_ic: &[
-		[
-			20,117,178,144,199,86,40,28,139,72,149,150,221,34,220,185,112,196,247,204,36,86,0,111,2,39,252,64,163,136,158,72,
-			14,116,172,32,18,129,193,47,179,128,153,220,86,97,240,34,171,181,104,5,72,194,175,108,20,184,146,61,171,238,127,250,
-		],
-		[
-			32,83,181,76,227,121,250,47,99,98,192,148,197,165,111,76,36,158,155,160,246,115,203,238,253,254,139,250,176,51,162,229,
-			26,203,227,24,200,114,78,250,92,69,180,67,48,84,214,122,58,41,134,87,194,168,123,165,103,228,149,242,190,208,204,40,
-		],
-		[
-			42,185,115,152,171,110,26,235,67,184,206,136,164,59,136,168,243,217,252,214,161,247,124,225,28,54,89,216,180,178,156,212,
-			37,219,158,110,208,228,240,100,139,45,98,39,5,120,91,68,178,248,209,142,68,44,161,99,67,228,197,219,137,238,36,106,
-		],
-		[
-			10,61,208,60,178,139,175,144,164,251,164,51,41,242,227,193,62,91,83,250,34,163,155,251,97,92,68,187,170,230,204,186,
-			15,74,29,129,199,37,101,163,127,57,111,6,135,106,61,37,67,103,196,228,27,134,237,126,248,183,26,110,106,124,86,132,
-		],
-		[
-			22,163,116,25,68,151,95,1,98,83,99,172,107,98,128,76,213,51,155,160,104,61,216,3,114,144,56,172,146,235,85,124,
-			19,251,92,184,124,129,6,66,241,98,69,167,26,12,193,77,162,141,144,139,106,118,227,121,247,220,23,4,243,133,14,113,
-		],
-	]
-};
