+++ anchor/programs/darklake/src/instructions/mod.rs
@@ -1,7 +1,6 @@
 pub mod initialize_pool;
 pub mod add_liquidity;
 pub mod confidential_swap;
-pub mod confidential_swap_exp;
 pub mod remove_liquidity;
 pub mod upgrade_pool;
 
@@ -9,5 +8,4 @@ pub use remove_liquidity::*;
 pub use initialize_pool::*;
 pub use add_liquidity::*;
 pub use confidential_swap::*;
-pub use confidential_swap_exp::*;
 pub use upgrade_pool::*;
