--- circuits/division_test.circom
@@ -1,5 +1,0 @@
-pragma circom 2.0.0;
-
-include "division.circom";
-
-component main = ReciprocalDivision(252);
