+++ circuits/swap.circom
@@ -1,3 +1,4 @@
+
 pragma circom 2.0.0;
 
 include "node_modules/circomlib/circuits/comparators.circom";
@@ -7,13 +8,18 @@ include "division.circom";
 
 template ZKConstantProductAMM() {
     // Private inputs
-    signal input privateInputAmount;
+    // TODO: for MVP this is equal to exact amountReceived
+    // Later improved to be immune to front-running
     signal input privateMinReceived;
 
     // Public inputs
-    signal input publicBalanceX;
-    signal input publicBalanceY;
+    // supplied by the user when swapping
+    signal input inputAmount;
     signal input isSwapXtoY; // 1 if swapping X to Y, 0 if swapping Y to X
+
+    // verified in the pool when swapping
+    signal input currentReserveX;
+    signal input currentReserveY;
 
     // Outputs
     signal output newBalanceX;
@@ -21,23 +27,23 @@ template ZKConstantProductAMM() {
     signal output amountReceived;
 
     // Calculate constant product
-    signal constantProduct <== publicBalanceX * publicBalanceY;
+    signal constantProduct <== currentReserveX * currentReserveY;
 
     // Determine swap direction and calculate amounts
     component muxInput = Mux1();
-    muxInput.c[0] <== publicBalanceY;
-    muxInput.c[1] <== publicBalanceX;
+    muxInput.c[0] <== currentReserveY;
+    muxInput.c[1] <== currentReserveX;
     muxInput.s <== isSwapXtoY;
     signal inputBalance <== muxInput.out;
 
     component muxOutput = Mux1();
-    muxOutput.c[0] <== publicBalanceX;
-    muxOutput.c[1] <== publicBalanceY;
+    muxOutput.c[0] <== currentReserveX;
+    muxOutput.c[1] <== currentReserveY;
     muxOutput.s <== isSwapXtoY;
     signal outputBalance <== muxOutput.out;
 
     // Calculate new input balance
-    signal newInputBalance <== inputBalance + privateInputAmount;
+    signal newInputBalance <== inputBalance + inputAmount;
 
     // Calculate new output balance (y = k / x)
     component division = ReciprocalDivision(252);
@@ -47,13 +53,15 @@ template ZKConstantProductAMM() {
 
     // Assign new balances
     signal intermediate1 <== (1 - isSwapXtoY) * newOutputBalance;
+    // TODO: needs further investigation how to handle newBalanceX and newBalanceY
     newBalanceX <== isSwapXtoY * newInputBalance + intermediate1;
     signal intermediate2 <== (1 - isSwapXtoY) * newInputBalance;
     newBalanceY <== isSwapXtoY * newOutputBalance + intermediate2;
 
     // Calculate amount received
-    signal intermediate3 <== isSwapXtoY * (publicBalanceY - newOutputBalance);
-    signal intermediate4 <== (1 - isSwapXtoY) * (publicBalanceX - newOutputBalance);
+    signal intermediate3 <== isSwapXtoY * (currentReserveY - newOutputBalance);
+    signal intermediate4 <== (1 - isSwapXtoY) * (currentReserveX - newOutputBalance);
+
     amountReceived <== intermediate3 + intermediate4;
 
     // Verify minimum received amount
@@ -64,7 +72,7 @@ template ZKConstantProductAMM() {
 
     // Range check for private inputs
     component privateInputAmountCheck = Num2Bits(252);
-    privateInputAmountCheck.in <== privateInputAmount;
+    privateInputAmountCheck.in <== inputAmount;
 
     component privateMinReceivedCheck = Num2Bits(252);
     privateMinReceivedCheck.in <== privateMinReceived;
@@ -84,4 +92,4 @@ template ZKConstantProductAMM() {
     // newBalanceX * newBalanceY === constantProduct;
 }
 
-component main = ZKConstantProductAMM();
+component main {public [inputAmount,currentReserveX,currentReserveY,isSwapXtoY]} = ZKConstantProductAMM();
