--- circuitsV2/verification_key.json
@@ -1,109 +1,0 @@
-{
- "protocol": "groth16",
- "curve": "bn128",
- "nPublic": 4,
- "vk_alpha_1": [
-  "17945886941149807785136888824699095888710946885985605235018534964045148744236",
-  "6672650003944701844988342268143667900627241467359683143666636955044131727770",
-  "1"
- ],
- "vk_beta_2": [
-  [
-   "16434015757509914576149482027209507932068451579009253731197558912953921393741",
-   "3861172734304167512316643766271714823130503426807563061805783228554090288568"
-  ],
-  [
-   "19185257365978393132255981370174504742887451384681518791896354286764408699209",
-   "8774872967233755841634726228819045728234131063396866816428548976214561213588"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_gamma_2": [
-  [
-   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-  ],
-  [
-   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_delta_2": [
-  [
-   "8895047992907908171386528853915283556533396787131907513418607648504580115902",
-   "9191359954033378577108450356665132303187113132269926092273691940224374501172"
-  ],
-  [
-   "7439720906295074628795762003901584526137600040812721784275050690647553839216",
-   "5994829828051775128113510759613806169208178603442106692297837705249460560690"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_alphabeta_12": [
-  [
-   [
-    "1476611972639559638905451864292408498640176089895511819637764979588588620760",
-    "2541832739063950726418360538944297759779086395928377178794610993317240925884"
-   ],
-   [
-    "7136129908334692003917878624045675248367102771072716290218522027981790730298",
-    "6924745163134522943076039723955574238706960413765999808275074587997984447596"
-   ],
-   [
-    "19374417280315230583681571836291073055325018889491726758676521288832816348834",
-    "1577005885089491649982077533519096338488869764736059715467780222115620290913"
-   ]
-  ],
-  [
-   [
-    "3987044168076989526980204528224224762270502645838967057331322701878496252799",
-    "535498105363742536496111489761057794112963112949042368841416854838900361343"
-   ],
-   [
-    "8628355751500602068350514081662682226455348744522588128455087940623103469516",
-    "16691909262447853762538053292419021731172050202753016374361419797111833004583"
-   ],
-   [
-    "13799687291702116204471160232029335145735678213223555823119202871810326597680",
-    "10906618023952551917662881659672368778009842219195400569090746132810355924270"
-   ]
-  ]
- ],
- "IC": [
-  [
-   "9254210492319036370795150144040222205048866769942282154446075457678273453640",
-   "6538522104718962555725459523307271983108047175053672844799489516397275283450",
-   "1"
-  ],
-  [
-   "14621910750041914876389162603175783400936852270327347738606005107831015187173",
-   "12120371381883861712507427846248036421336377387709523925621000243752649935912",
-   "1"
-  ],
-  [
-   "19324804164276660242203535385036504544043613919512593452497204543127234452692",
-   "17123608368283385556721144408178804476029111400585642183550475884205427270762",
-   "1"
-  ],
-  [
-   "4632343356424129116619853977352228657452904425521003682860101279731008588986",
-   "6915643060992303540999081866671590615359876944203744443880910727246046320260",
-   "1"
-  ],
-  [
-   "10239680024189137872403556999158039654731754022051622617113583230124248749436",
-   "9038062670747329084115128468385135022748478484560709161433344766226477682289",
-   "1"
-  ]
- ]
-}
